requirements.txt
aiohttp>=3.12.15
python-dotenv>=1.1.1
websockets>=15.0.1

